Feat: Github action e2e Testing and bash script
Features a new e2e CI based on a bash script. the script makes some
tests to verify that the operator is working as intended. It does not go
into really deep tests, this is a way to simulate how a classic user
would use kotary. Mainly used to verify that the operator is working
fine.